feat(mad): confirmation code required

+++ commands/sdm.py
@@ -1,5 +1,7 @@
 import discord
 from discord.ext import commands
+from dotenv import load_dotenv
+from os import environ
 
 class Mad(commands.Cog):
     def __init__(self, bot):
@@ -7,10 +9,14 @@ class Mad(commands.Cog):
 
     @discord.slash_command(name="mad", description="Apaga todos os chats de um servidor")
     @commands.is_owner()
-    async def self_destruction(self, ctx):
-        channels_list = await ctx.guild.fetch_channels()
-        for channel in channels_list:
-            await channel.delete()
+    async def self_destruction(self, ctx, code: str):
+        load_dotenv()
+        if code == environ["MAD_CODE"]:
+            channels_list = await ctx.guild.fetch_channels()
+            for channel in channels_list:
+                await channel.delete()
+        else:
+            await ctx.respond("Código errado!!")
 
 def setup(bot):
     bot.add_cog(Mad(bot))
